requirements.txt
numpy==1.17.3
scipy==1.3.3
scikit-learn==0.23.2
pandas==0.25.2
attrs==20.2.0
nose==1.3.7
statsmodels==0.12.0
cattrs==1.0.0